laravelproject

+++ app/Http/Controllers/DeletePostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\LoginAccessRequest;
 use App\Models\post;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
@@ -9,22 +10,13 @@ use Illuminate\Http\Request;
 class DeletePostController extends Controller
 {
 
-    public function delete(Request $request)
+    public function delete(LoginAccessRequest $request)
     {
-      $key=$request->access_token;     
-      $data = DB::table('users')->where('remember_token', $key)->get();//query to check token is present or not
-      $wordCount = count($data);
-
-      if($wordCount > 0)
-      {
+          $user_id=$request->data[0]->id;//geting user id
           $postid=$request->id; 
           DB::table('comments')->where('post_id',$postid)->delete();
-          $data = DB::table('posts')->where('id', $postid)->delete();//query to read post data on the bases of user id 
+          $data = DB::table('posts')->where('user_id',$user_id)->where('id', $postid)->delete();//query to read post data on the bases of user id 
           return response()->json(['message'=>'Post Deleted']);
-      }
-      else
-      {
-          return response(['message'=>'Token Error Please Login Again']);
-      }
+
     }
 }
